Refactor elevate command and move kubeconfig func to kubeconfig module
