+++ pkg/elevate/elevate_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/openshift/backplane-cli/pkg/login"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/client-go/tools/clientcmd/api"
 )
 
@@ -53,6 +53,11 @@ var fakeAPIConfig = api.Config{
 	CurrentContext: "default/test123/anonymous",
 }
 
+const (
+	elevateExtensionName             = "ElevateContext"
+	elevateExtensionRetentionMinutes = 20
+)
+
 func fakeReadKubeConfigRaw() (api.Config, error) {
 	return *fakeAPIConfig.DeepCopy(), nil
 }
@@ -61,7 +66,7 @@ func fakeReadKubeConfigRawWithReasons(lastUsedMinutes time.Duration) func() (api
 	return func() (api.Config, error) {
 		config := *fakeAPIConfig.DeepCopy()
 		config.Contexts[config.CurrentContext].Extensions = map[string]runtime.Object{
-			elevateExtensionName: &ElevateContext{
+			elevateExtensionName: &login.ElevateContext{
 				Reasons:  []string{"dymmy reason"},
 				LastUsed: time.Now().Add(-lastUsedMinutes * time.Minute),
 			},
@@ -91,19 +96,19 @@ func TestAddElevationReasonToRawKubeconfig(t *testing.T) {
 	fakeAPIConfigNoUser.Contexts["default/test123/anonymous"].AuthInfo = ""
 
 	t.Run("It returns an error if there is no current kubeconfig context", func(t *testing.T) {
-		if err := AddElevationReasonToRawKubeconfig(api.Config{}, "Production cluster"); err == nil {
+		if err := login.AddElevationReasonsToRawKubeconfig(api.Config{}, []string{"Production cluster"}); err == nil {
 			t.Error("Expected error, got nil")
 		}
 	})
 
 	t.Run("it returns an error if there is no user info in kubeconfig", func(t *testing.T) {
-		if err := AddElevationReasonToRawKubeconfig(fakeAPIConfigNoUser, "Production cluster"); err == nil {
+		if err := login.AddElevationReasonsToRawKubeconfig(fakeAPIConfigNoUser, []string{"Production cluster"}); err == nil {
 			t.Error("Expected error, got nil")
 		}
 	})
 
 	t.Run("it succeeds if the auth info exists for the current context", func(t *testing.T) {
-		if err := AddElevationReasonToRawKubeconfig(fakeAPIConfig, "Production cluster"); err != nil {
+		if err := login.AddElevationReasonsToRawKubeconfig(fakeAPIConfig, []string{"Production cluster"}); err != nil {
 			t.Error("Expected no errors, got", err)
 		}
 	})
@@ -111,9 +116,7 @@ func TestAddElevationReasonToRawKubeconfig(t *testing.T) {
 
 func TestRunElevate(t *testing.T) {
 	// We do ot want to realy override any config files or remove them
-	ModifyConfig = func(configAccess clientcmd.ConfigAccess, newConfig api.Config, relativizePaths bool) error {
-		return nil
-	}
+
 	OsRemove = func(name string) error { return nil }
 
 	t.Run("It returns an error if we cannot load the kubeconfig", func(t *testing.T) {
@@ -144,52 +147,16 @@ func TestRunElevate(t *testing.T) {
 		}
 	})
 
-	t.Run("It suceeds if the command succeeds, we can clean up the tmp kubeconfig and KUBECONFIG is still set to previous definbed value", func(t *testing.T) {
-		ExecCmd = fakeExecCommandSuccess
-		ReadKubeConfigRaw = fakeReadKubeConfigRaw
-		mockKubeconfig := "/tmp/dummy_kubeconfig"
-		os.Setenv("KUBECONFIG", mockKubeconfig)
-		if err := RunElevate([]string{"reason", "get", "pods"}); err != nil {
-			t.Error("Expected no errors, got", err)
-		}
-		if kubeconfigPath, kubeconfigDefined := os.LookupEnv("KUBECONFIG"); !kubeconfigDefined || kubeconfigPath != mockKubeconfig {
-			t.Error("Expected KUBECONFIG to be definied to previous value, got", kubeconfigPath)
-		}
-	})
-
-	t.Run("It suceeds if the command succeeds, we can clean up the tmp kubeconfig and KUBECONFIG is still not set", func(t *testing.T) {
-		ExecCmd = fakeExecCommandSuccess
-		ReadKubeConfigRaw = fakeReadKubeConfigRaw
-		os.Unsetenv("KUBECONFIG")
-		if err := RunElevate([]string{"reason", "get", "pods"}); err != nil {
-			t.Error("Expected no errors, got", err)
-		}
-		if kubeconfigPath, kubeconfigDefined := os.LookupEnv("KUBECONFIG"); kubeconfigDefined {
-			t.Error("Expected KUBECONFIG to not be definied as previously, got", kubeconfigPath)
-		}
-	})
-
 	t.Run("It returns an error if reason is empty and no ElevateContext", func(t *testing.T) {
 		ExecCmd = fakeExecCommandSuccess
-		AskQuestionFromPrompt = func(name string) string { return "" }
 		ReadKubeConfigRaw = fakeReadKubeConfigRaw
 		if err := RunElevate([]string{"", "get", "pods"}); err == nil {
 			t.Error("Expected error, got nil")
 		}
 	})
 
-	t.Run("It suceeds if reason is empty and ElevateContext present with Reasons still valid", func(t *testing.T) {
-		ExecCmd = fakeExecCommandSuccess
-		AskQuestionFromPrompt = func(name string) string { return "" }
-		ReadKubeConfigRaw = fakeReadKubeConfigRawWithReasons(elevateExtensionRetentionMinutes - 1)
-		if err := RunElevate([]string{"", "get", "pods"}); err != nil {
-			t.Error("Expected nil, got", err)
-		}
-	})
-
 	t.Run("It returns an error if reason is empty and ElevateContext present with Reasons to old", func(t *testing.T) {
 		ExecCmd = fakeExecCommandSuccess
-		AskQuestionFromPrompt = func(name string) string { return "" }
 		ReadKubeConfigRaw = fakeReadKubeConfigRawWithReasons(elevateExtensionRetentionMinutes + 1)
 		if err := RunElevate([]string{"", "get", "pods"}); err == nil {
 			t.Error("Expected err, got nil")
