+++ pkg/login/kubeConfig.go
@@ -6,18 +6,54 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"time"
 
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/client-go/tools/clientcmd/api"
 
 	logger "github.com/sirupsen/logrus"
 
 	"github.com/openshift/backplane-cli/pkg/info"
+	"github.com/openshift/backplane-cli/pkg/utils"
 )
 
 var (
 	kubeConfigBasePath string
 )
+
+const (
+	elevateExtensionName             = "ElevateContext"
+	elevateExtensionRetentionMinutes = 20
+)
+
+type ElevateContext struct {
+	Reasons  []string  `json:"reasons"`
+	LastUsed time.Time `json:"lastUsed"`
+}
+
+///////////////////////////////////////////////////////////////////////
+// runtime.Object interface func implementation for ElevateContext type
+
+// DeepCopyObject creates a deep copy of the ElevateContext.
+func (r *ElevateContext) DeepCopyObject() runtime.Object {
+	return &ElevateContext{
+		Reasons:  append([]string(nil), r.Reasons...),
+		LastUsed: r.LastUsed,
+	}
+}
+
+// GetObjectKind returns the schema.GroupVersionKind of the object.
+func (r *ElevateContext) GetObjectKind() schema.ObjectKind {
+	// return schema.EmptyObjectKind
+	return &runtime.TypeMeta{
+		Kind:       "ElevateContext",
+		APIVersion: "v1",
+	}
+}
+
+///////////////////////////////////////////////////////////////////////
 
 // CreateClusterKubeConfig creates cluster specific kube config based on a cluster ID
 func CreateClusterKubeConfig(clusterID string, kubeConfig api.Config) (string, error) {
@@ -134,4 +170,90 @@ func getKubeConfigBasePath() (string, error) {
 func SetKubeConfigBasePath(basePath string) error {
 	kubeConfigBasePath = basePath
 	return nil
+}
+
+// in some cases (mainly when config is created from json) the "ElevateContext Extension" is created as runtime.Unknow object
+// instead of the desired ElevateContext, so we need to Unmarshal the raw definition in that case
+func GetElevateContextReasons(config api.Config) []string {
+	if currentContext := config.Contexts[config.CurrentContext]; currentContext != nil {
+		var elevateContext *ElevateContext
+		var ok bool
+		if object := currentContext.Extensions[elevateExtensionName]; object != nil {
+			//Let's first try to cast the extension object in ElevateContext
+			if elevateContext, ok = object.(*ElevateContext); !ok {
+				// and if it does not work, let's try cast the extension object in Unknown
+				if unknownObject, ok := object.(*runtime.Unknown); ok {
+					// and unmarshal the unknown raw JSON string into the ElevateContext
+					_ = json.Unmarshal([]byte(unknownObject.Raw), &elevateContext)
+				}
+			}
+			// We should keep the stored ElevateContext only if it is still valid
+			if elevateContext != nil && time.Since(elevateContext.LastUsed) <= elevateExtensionRetentionMinutes*time.Minute {
+				return elevateContext.Reasons
+			}
+		}
+	}
+	return []string{}
+}
+
+func AddElevationReasonsToRawKubeconfig(config api.Config, elevationReasons []string) error {
+	logger.Debugln("Adding reason for backplane-cluster-admin elevation")
+	if config.Contexts[config.CurrentContext] == nil {
+		return errors.New("no current kubeconfig context")
+	}
+
+	currentCtxUsername := config.Contexts[config.CurrentContext].AuthInfo
+
+	if config.AuthInfos[currentCtxUsername] == nil {
+		return errors.New("no current user information")
+	}
+
+	if config.AuthInfos[currentCtxUsername].ImpersonateUserExtra == nil {
+		config.AuthInfos[currentCtxUsername].ImpersonateUserExtra = make(map[string][]string)
+	}
+
+	config.AuthInfos[currentCtxUsername].ImpersonateUserExtra["reason"] = elevationReasons
+	config.AuthInfos[currentCtxUsername].Impersonate = "backplane-cluster-admin"
+
+	return nil
+}
+
+func SaveElevateContextReasons(config api.Config, elevationReason string) ([]string, error) {
+	currentCtx := config.Contexts[config.CurrentContext]
+	if currentCtx == nil {
+		return nil, errors.New("no current kubeconfig context")
+	}
+
+	// let's first retrieve previous elevateContext if any, and add any provided reason.
+	elevationReasons := utils.AppendUniqNoneEmptyString(
+		GetElevateContextReasons(config),
+		elevationReason,
+	)
+
+	// if we still do not have reason, then let's try to have the reason from prompt
+	if len(elevationReasons) == 0 {
+		elevationReasons = utils.AppendUniqNoneEmptyString(
+			elevationReasons,
+			utils.AskQuestionFromPrompt(fmt.Sprintf("Please enter a reason for elevation, it will be stored in current context for %d minutes : ", elevateExtensionRetentionMinutes)),
+		)
+	}
+	// and raise an error if not possible
+	if len(elevationReasons) == 0 {
+		return nil, errors.New("please enter a reason for elevation")
+	}
+
+	// Store the ElevateContext in config current context Extensions map
+	if currentCtx.Extensions == nil {
+		currentCtx.Extensions = map[string]runtime.Object{}
+	}
+	currentCtx.Extensions[elevateExtensionName] = &ElevateContext{
+		Reasons:  elevationReasons,
+		LastUsed: time.Now(),
+	}
+
+	// Save the config to default path.
+	configAccess := clientcmd.NewDefaultPathOptions()
+	err := clientcmd.ModifyConfig(configAccess, config, true)
+
+	return elevationReasons, err
 }
